private functions referenced by closures make Erick something something...
the OX.Mixins private function getURI was inappropriately caching the pubSubURI of the first
caller into it onto the function itself.  When we extend OX objects the functions are
not copied, they are referenced, we just point the prototype of the new object to an instance of the old
object.  This obviously means we can't cache a value on an abstract base class function because
then any child classes will all return the same value to that function call - the value being whoever
called it first.  This all led to jidHandlers not be saved properly on the OX.Connection which then
meant you could only register a single service with OX... not good.